Ensure Gifsicle is used for image/gif tranformations.
This commit ensures the Gifsicle library is available and thus is
picked up by the express-processimage pipeline. Add tests.

Cover the of a rotation of zero being supplied for GIFs. While the
pipeline correctly skips doing work in this case, it unfortunately
returns something that is not a stream to we special case this and
instead return a PassThrough stream so callers are unaffected.