Pass through unconverted string args and handle more errors.
With the port to impro some glue was added to continue to support
shorthand string arguments for "resize", but this properly handle
string arguments that do not need conversion. Address this and
cover the situation with tests.

While here ensure that any exceptions caused by conversions done
on string arguments will be wrapped wrapped in an error with a
clear message which is then propogated.