Implement maxInputPixels and maxOutputPixels.
The impro library supports placing limits on the size of processed
images, the use-case being to avoid committing too many resources
to processing huge images which could become problematic if such
processing is done dynamically in server code responding to requests.

Allow these options to be passed within an imageOptions object. This
also means they can be set dynamically based on the detected image
metadata via the postProcessImageOptions callback.

Note that a special formulation of the tests was required for these
cases - it seems that a .then() attached to an expect call using the
"to call the callback without error" assertion can mean handlers get
attached to the output stream too late for the first error event to
be registered in the case of an output pixels error.